🛡️ Guardian (JULES): Consolidate utilities into canonical modules
WHAT:
- Merged `extractVariables` and `ExtractedVariable` into `src/lib/variable-detection.ts`
- Removed inline `slugify` from `src/pages/api/mcp.ts`
- Updated imports in `src/pages/api/mcp.ts` to use canonical modules

WHY:
- Identical implementations of utility functions causing maintenance burden
- Centralizing these functions creates a single source of truth

JULES CONTEXT:
- Checked .Jules/task-log.md: No Autonomous conflicts
- Session logged: .Jules/guardian/2026-07-09/
- Reference: .Jules/JULES.md

IMPACT:
- Reused canonical `slugify` logic
- Centralized `extractVariables`
- Zero duplicated utility definitions
- No breaking changes

VERIFICATION:
- All tests pass
- TypeScript check: ✓
- Linting: ✓
- JULES compliance: ✓
