+++ .Jules/guardian/2026-07-09/session-report.md
@@ -1,0 +1,7 @@
+## 2026-07-09 - [Consolidated variables and slugify]
+
+**Target:** src/pages/api/mcp.ts, src/lib/variable-detection.ts
+**Learning:** Found multiple instances of inline utility implementations like \`slugify\` and \`extractVariables\` that should belong in central shared utility files to avoid technical debt and inconsistencies.
+**Action:** Relocated \`extractVariables\` and \`ExtractedVariable\` to \`src/lib/variable-detection.ts\`, removed inline \`slugify\`, and updated \`src/pages/api/mcp.ts\` imports.
+**JULES Check:** Verified no Autonomous task conflicts in .Jules/task-log.md.
+**Conflicts Avoided:** None.
