+++ .Jules/task-log.md
@@ -1,0 +1,7 @@
+# Task Log
+
+## 2026-07-09 16:34 - [GUARDIAN] Session started
+
+- Directory: .Jules/guardian/2026-07-09
+- Phase: PRE-FLIGHT
+- JULES Check: COMPLETE
