+++ src/pages/api/mcp.ts
@@ -11,6 +11,8 @@ import { z } from "zod";
 import { db } from "@/lib/db";
 import { isValidApiKeyFormat } from "@/lib/api-key";
 import { improvePrompt } from "@/lib/ai/improve-prompt";
+import { slugify } from "@/lib/slug";
+import { extractVariables, type ExtractedVariable } from "@/lib/variable-detection";
 import {
   parseSkillFiles,
   serializeSkillFiles,
@@ -41,20 +43,6 @@ async function authenticateApiKey(apiKey: string | null): Promise<AuthenticatedU
   return user;
 }
 
-interface ExtractedVariable {
-  name: string;
-  defaultValue?: string;
-}
-
-function slugify(text: string): string {
-  return text
-    .toLowerCase()
-    .trim()
-    .replace(/[^\w\s-]/g, "")
-    .replace(/[\s_-]+/g, "-")
-    .replace(/^-+|-+$/g, "");
-}
-
 /**
  * Get the prompt name/slug for MCP.
  * Priority: slug > slugify(title) > id
@@ -64,25 +52,6 @@ function getPromptName(prompt: { id: string; slug?: string | null; title: string
   const titleSlug = slugify(prompt.title);
   if (titleSlug) return titleSlug;
   return prompt.id;
-}
-
-function extractVariables(content: string): ExtractedVariable[] {
-  // Format: ${variableName} or ${variableName:default}
-  const regex = /\$\{([a-zA-Z_][a-zA-Z0-9_\s]*?)(?::([^}]*))?\}/g;
-  const variables: ExtractedVariable[] = [];
-  const seen = new Set<string>();
-  let match;
-  while ((match = regex.exec(content)) !== null) {
-    const name = match[1].trim();
-    if (!seen.has(name)) {
-      seen.add(name);
-      variables.push({
-        name,
-        defaultValue: match[2]?.trim(),
-      });
-    }
-  }
-  return variables;
 }
 
 export const config = {
